Fix incorrect typing in imgui wgpu impl

+++ AxEng/backends/imgui_impl_wgpu.cpp
@@ -441,8 +441,8 @@ static void ImGui_ImplWGPU_SetupRenderState(ImDrawData* draw_data, WGPURenderPas
     }
 
     // Setup viewport
-    int fb_width = (int)(draw_data->DisplaySize.x * draw_data->FramebufferScale.x);
-    int fb_height = (int)(draw_data->DisplaySize.y * draw_data->FramebufferScale.y);
+    float fb_width = (float)(draw_data->DisplaySize.x * draw_data->FramebufferScale.x);
+    float fb_height = (float)(draw_data->DisplaySize.y * draw_data->FramebufferScale.y);
     wgpuRenderPassEncoderSetViewport(ctx, 0, 0, fb_width, fb_height, 0, 1);
 
     // Bind shader and vertex buffers
